+++ docs/CHANGELOG.md
@@ -6,6 +6,7 @@
 * Storage (Google Drive) settings have moved to the Settings page.
 * You can configure a custom item sorting method from the Settings page.
 * Improved display of the account selection dropdown.
+* Mobile on portrait mode will be able to set the number of inventory columns (the icon size will be resized to accomodate)
 
 # 4.36.1